scripts/requirements.txt
lxml
requests
pyscopg2
python-dotenv